+++ src/enums/state_enum.py
@@ -1,0 +1,8 @@
+
+class DeviceState:
+    """
+    Enum representing any device state.
+    """
+    ON = 1
+    OFF = 2
+    IN_ACTION = 3
